chore: move toggle component
