+++ packages/ui/index.ts
@@ -14,3 +14,4 @@ export { PageLoading } from './PageLoading';
 export { Spinner } from './Spinner';
 export { default as TabButton } from './TabButton';
 export { TextArea } from './TextArea';
+export { Toggle } from './Toggle';
